app/utils/heatmap.py
import os
import torch
import numpy as np
from PIL import Image
import matplotlib.pyplot as plt
from typing import Dict, Any, Optional


def generate_heatmap(image_path: str, image_tensor: torch.Tensor, prediction_results: Dict[str, Any]) -> Optional[str]:
    """Generate a heatmap visualization for the model's prediction
    
    This is a simplified implementation of Grad-CAM. In a real implementation,
    you would need to hook into the model's forward pass to get the feature maps
    and gradients. This function simulates that process with a placeholder.
    
    Args:
        image_path: Path to the original image
        image_tensor: Preprocessed image tensor used for inference
        prediction_results: Results from the model inference
        
    Returns:
        Path to the generated heatmap image, or None if generation failed
    """
    try:
        # Create temp directory if it doesn't exist
        os.makedirs("temp", exist_ok=True)
        
        # Get the upload ID from the image path
        upload_id = os.path.splitext(os.path.basename(image_path))[0]
        
        # Create heatmap path
        heatmap_path = os.path.join("temp", f"{upload_id}_heatmap.jpg")
        
        # Load the original image
        original_image = Image.open(image_path).convert("RGB")
        original_image = original_image.resize((224, 224))
        
        # In a real implementation, you would generate the actual heatmap here
        # using Grad-CAM or a similar technique. For now, we'll create a simulated heatmap.
        
        # Create a simulated heatmap (random for demonstration)
        # In a real implementation, this would be based on the model's attention
        heatmap = np.random.rand(224, 224)
        
        # Apply a gaussian filter to make it look more realistic
        from scipy.ndimage import gaussian_filter
        heatmap = gaussian_filter(heatmap, sigma=10)
        
        # Normalize the heatmap
        heatmap = (heatmap - heatmap.min()) / (heatmap.max() - heatmap.min())
        
        # Create a color map
        plt.figure(figsize=(10, 8))
        plt.subplot(1, 3, 1)
        plt.imshow(original_image)
        plt.title("Original Image")
        plt.axis('off')
        
        plt.subplot(1, 3, 2)
        plt.imshow(heatmap, cmap='jet')
        plt.title("Heatmap")
        plt.axis('off')
        
        plt.subplot(1, 3, 3)
        plt.imshow(original_image)
        plt.imshow(heatmap, cmap='jet', alpha=0.5)
        plt.title(f"Prediction: {prediction_results['label']} ({prediction_results['confidence']:.1f}%)")
        plt.axis('off')
        
        plt.tight_layout()
        plt.savefig(heatmap_path)
        plt.close()
        
        return heatmap_path
    
    except Exception as e:
        print(f"Error generating heatmap: {e}")
        return None