requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
onnxruntime==1.16.0
torchvision==0.16.0
supabase==1.0.3
psycopg2-binary==2.9.9
python-multipart==0.0.6
Pillow==10.0.1
python-dotenv==1.0.0
numpy==1.24.3
requests==2.31.0
pydantic==2.4.2